Clean up init code
