--- src/KeycloakContext.ts
@@ -1,4 +1,0 @@
-import { createContext } from 'react'
-import Keycloak from 'keycloak-js'
-
-export const KeycloakContext = createContext<Keycloak>(new Keycloak())
