+++ src/components/ToolBar/DataMenu/LoadWorkspaceMenuItem.tsx
@@ -1,5 +1,5 @@
 import { Box, MenuItem, Tooltip } from '@mui/material'
-import { KeycloakContext } from '../../../KeycloakContext'
+import { KeycloakContext } from '../../../init/keycloak'
 import { ReactElement, useContext, useState } from 'react'
 import { BaseMenuProps } from '../BaseMenuProps'
 import { LoadWorkspaceDialog } from './LoadWorkspaceDialog'
